Scrum 119 post processing ramió

+++ Survival/Assets/Scripts/Personatges/Personatge.cs
@@ -8,9 +8,9 @@ public abstract class Personatge : MonoBehaviour
     public event Action QuanCanviVida;
 
     // Propietats
-    public abstract int VidaActual { get; }
-    public abstract int VidaMaxima { get; }
-    public abstract int Dany { get; }
+    public abstract float VidaActual { get; }
+    public abstract float VidaMaxima { get; }
+    public abstract float Dany { get; }
     public abstract float ForcaKnockback { get; }
     public abstract bool EstaAtacant();
 
@@ -18,7 +18,7 @@ public abstract class Personatge : MonoBehaviour
 
     protected abstract void Start();
     
-    public abstract void DecrementarVida(int quantitat, string font = "");
+    public abstract void DecrementarVida(float quantitat, string font = "");
 
     public abstract IEnumerator Morir();
